data_science/reddit_sentiment/reddit/requirements.txt
praw==7.6.1
pymongo==4.3.3
python-dotenv==0.21.0
PyYAML==6.0
